fix: made changes to overview ui
