+++ components/asset-allocation/Overview.jsx
@@ -37,39 +37,33 @@ const Overview = ({ totalAssetManagers, totalCap, topFive }) => {
         </div>
 
         {/* Exposure */}
+       
         <div className="my-12">
-          <div className="flex items-center justify-between border-b pb-4">
-            <p className="">Exposure</p>
-            <div className="flex  items-center gap-36 justify-between">
-              <p>Top 5 by allocation size</p>
-              <span className="flex gap-9">
-                <div
-                  data-tooltip-target="tooltip-default"
-                  className="relative inline-flex items-center justify-center w-10 h-10 overflow-hidden bg-black rounded-full "
-                >
-                  <span className="font-medium text-white ">?</span>
-                  <div
-                    id="tooltip-default"
-                    role="tooltip"
-                    className="absolute z-10 invisible inline-block px-3 py-2 text-sm font-medium text-white bg-gray-900 rounded-lg shadow-sm opacity-0 tooltip dark:bg-gray-700"
-                  >
-                    Tooltip content
-                    <div className="tooltip-arrow" data-popper-arrow></div>
-                  </div>
-                </div>
-                <div className="relative inline-flex items-center justify-center w-10 h-10 overflow-hidden bg-black rounded-full ">
-                  <span className="font-medium text-white ">?</span>
-                </div>
-                <div className="relative inline-flex items-center justify-center w-10 h-10 overflow-hidden bg-black rounded-full ">
-                  <span className="font-medium text-white ">?</span>
-                </div>
-              </span>
+      <div className="grid grid-cols-2 items-center border-b  pb-4">
+        <div className="text-sm font-bold">Exposure</div>
+        <div className="grid grid-cols-2 items-center gap-36 ">
+          <div className="capitaliuze bg-gray-100 text-sm font-bold">Top 5 by allocation size</div>
+          <div className="grid grid-cols-3">
+            <div data-tooltip-target="tooltip-default" className="grid h-8 w-8 place-items-center place-self-end rounded-full bg-black">
+              <span className="font-medium text-white">?</span>
+              <div id="tooltip-default" role="tooltip" className="tooltip invisible absolute z-10 inline-block rounded-lg bg-gray-900 px-3 py-2 text-sm font-medium text-white opacity-0 shadow-sm dark:bg-gray-700">
+                Tooltip content
+                <div className="tooltip-arrow" data-popper-arrow></div>
+              </div>
+            </div>
+            <div className="grid h-8 w-8 place-items-center place-self-end rounded-full bg-black">
+              <span className="font-medium text-white">?</span>
+            </div>
+            <div className="grid h-8 w-8 place-items-center place-self-end rounded-full bg-black">
+              <span className="font-medium text-white">?</span>
             </div>
           </div>
         </div>
+      </div>
+    </div>
 
         {/* Pie Chart & Legend */}
-        <div className="flex items-center justify-between">
+        {/* <div className="flex items-center justify-between">
           <div className=" h-96  mx-auto">
             <PieChart datum={topFive} />
           </div>
@@ -85,7 +79,7 @@ const Overview = ({ totalAssetManagers, totalCap, topFive }) => {
                     <p className="w-12 ">{formatCompact(top.total)}</p>
                     <p className="w-12 ">
                       {Math.ceil((parseFloat(top.total) / totalCap) * 100)}%
-                      {/* 100% */}
+                     
                     </p>
                     {Math.ceil((parseFloat(top.total) / totalCap) * 100) >
                     50 ? (
@@ -98,10 +92,37 @@ const Overview = ({ totalAssetManagers, totalCap, topFive }) => {
               );
             })}
           </div>
-        </div>
+        </div> */}
+
+<div className="grid grid-cols-2  items-center justify-between">
+      <div className="mx-auto grid place-content-center ">
+        <PieChart datum={topFive} />
       </div>
+
+     <div>
+      {topFive?.map((top) => {
+              return (
+                <div className=" grid grid-cols-2 gap-36 border-b  py-4" key={top.name}>
+                  <p className="">{top.name}</p>
+        <div className="grid grid-cols-3 items-center ">
+          <p className="place-self-end">{formatCompact(top.total)}</p>
+          <p className="place-self-end">{Math.ceil((parseFloat(top.total) / totalCap) * 100)}%</p>
+          
+          {Math.ceil((parseFloat(top.total) / totalCap) * 100) >
+                    50 ? (
+                      <p className="h-3 w-3 justify-self-end rounded-full bg-red-500"></p>
+                    ) : (
+                      <p className="h-3 w-3 justify-self-end rounded-full bg-green-500"></p>
+                    )}
+        </div>
+        </div>
+              )})}
+     </div>
+
     </div>
-  );
-};
+    </div>
+    </div>
+  )
+}
 
 export default Overview;
