+++ components/asset-allocation/dummy.jsx
@@ -1,25 +1,14 @@
 import axios from "axios";
 import { Button, Label, TextInput, Textarea, Modal } from "flowbite-react";
-import { Dialog, Transition } from '@headlessui/react'
-import React, { useRef, useState, Fragment, } from "react";
+import React, { useRef, useState } from "react";
 import { useRouter } from "next/router";
 import { ASSETMANAGERS } from "../../lib/api";
 
 export default function Buttons() {
-  let [isOpen, setIsOpen] = useState(true)
-
-  function closeModal() {
-    setIsOpen(false)
-  }
-
-  function openModal() {
-    setIsOpen(true)
-  }
-
   const router = useRouter();
   const { fundmanagerId } = router.query;
 
-  // const [openModal, setOpenModal] = useState(false);
+  const [openModal, setOpenModal] = useState(false);
   const [formData, setFormData] = useState({
     amount: "",
     description: "",
@@ -27,17 +16,14 @@ export default function Buttons() {
     // date: "",
     // maturityDate: "",
   });
-  
+  const rootRef = useRef(null);
 
-  function closeModal() {
-    setIsOpen(false)
-  }
-
-  function openModal() {
-    setIsOpen(true)
-  }
-
- 
+  const onClick = () => {
+    setOpenModal(true);
+  };
+  const onClose = () => {
+    setOpenModal(!true);
+  };
 
   const submitForm = async (e) => {
     e.preventDefault();
@@ -58,7 +44,6 @@ export default function Buttons() {
       createReminder(formData);
       console.log(response.data);
       onClose();
-      closeModal()
       router.reload();
     } catch (error) {
       console.log("error submiting form,", error);
@@ -96,56 +81,39 @@ export default function Buttons() {
       [name]: value,
     }));
   };
- 
+  const clickMe = () => {
+    console.log("clcicked me");
+  };
 
- 
+  const [show, setShow] = useState(false);
+  const handleChanges = (selectedDate) => {
+    console.log(selectedDate);
+  };
+  const handleClose = (state) => {
+    setShow(state);
+  };
 
   return (
     <>
       <Button
         color={"#2D7EC2"}
         className="flex items-end bg-[#2D7EC2] text-white"
-        onClick={openModal}
+        onClick={onClick}
         type="submit"
       >
         Add Transaction
       </Button>
 
-       <Transition appear show={isOpen} as={Fragment}>
-        <Dialog as="div" className="relative z-10" onClose={closeModal}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-black bg-opacity-25" />
-          </Transition.Child>
-
-          <div className="fixed inset-0 overflow-y-auto">
-            <div className="flex min-h-full items-center justify-center p-4 text-center">
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="opacity-0 scale-95"
-                enterTo="opacity-100 scale-100"
-                leave="ease-in duration-200"
-                leaveFrom="opacity-100 scale-100"
-                leaveTo="opacity-0 scale-95"
-              >
-                <Dialog.Panel className="w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
-                  <Dialog.Title
-                    as="h3"
-                    className="text-lg font-medium leading-6 text-gray-900 mb-3"
-                  >
-                   Add New Transaction
-                  </Dialog.Title>
-                  
-
-                  <form className="flex flex-col gap-4">
+      <div className=" mb-2 flex mx-auto items-center" ref={rootRef}>
+        <Modal
+          onClose={onClose}
+          show={openModal}
+          dismissible
+          root={rootRef.current ?? undefined}
+        >
+          <Modal.Header>Add New Transaction</Modal.Header>
+          <Modal.Body>
+            <form className="flex flex-col gap-4">
               <div>
                 <div className="mb-2 block">
                   <Label htmlFor="amount" value="Amount " />
@@ -161,7 +129,21 @@ export default function Buttons() {
                   onFocus={(e) => e.stopPropagation()}
                 />
               </div>
-              
+              {/* <div>
+                <div className="mb-2 block">
+                  <Label htmlFor="email1" value="Rate " />
+                </div>
+                <TextInput
+                  id="rate"
+                  name="rate"
+                  onChange={handleChange}
+                  value={formData.rate}
+                  placeholder="rate"
+                  required
+                  type="number"
+                  onFocus={(e) => e.stopPropagation()}
+                />
+              </div> */}
               <div>
                 <div className="mb-2 block">
                   <Label htmlFor="description" value="Description" />
@@ -177,30 +159,18 @@ export default function Buttons() {
                 />
               </div>
               
-              {/* <Button
+              <Button
                 color={"#2D7EC2"}
                 className="text-white bg-[#2D7EC2]"
                 type="submit"
                 onClick={submitForm}
               >
                 Submit
-              </Button> */}
-            </form>
-                  <div className="mt-4">
-                    <button
-                      type="button"
-                      className="inline-flex justify-center rounded-md border border-transparent bg-[#2D7EC2] px-4 py-2 text-sm font-medium text-white focus:outline-none focus-visible:ring-2 focus-visible:ring-blue-500 focus-visible:ring-offset-2"
-                      onClick={submitForm}
-                    >
-                      Submit
-                    </button>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
+              </Button>
+            </form>{" "}
+          </Modal.Body>
+        </Modal>
+      </div>
     </>
   );
 }
