+++ pages/asset-allocation/fundmanager/[fundmanagerId].js
@@ -8,6 +8,16 @@ import Header from "../../../components/header";
 import { ASSETMANAGERS } from "../../../lib/api";
 
 const formatCompact = (number) => {
+
+  let [isOpen, setIsOpen] = useState(true)
+
+  function closeModal() {
+    setIsOpen(false)
+  }
+
+  function openModal() {
+    setIsOpen(true)
+  }
   const formatter = new Intl.NumberFormat("en", {
     notation: "standard",
   });
